src-worddic: infinite loop if the end of anthy journal file is broken
Normally the end of anthy journal file, $CONFIG/last-record2_default.utf8,
should be '\n' when we check it with 'od' command.
But if that file is broken, the end of file would not be '\n'
and anthy_do_create_context() could go into an infinite loop.

The solution is to check EOF besides '\n' in record.c:read_1_token().

BUG=https://ja.osdn.net/projects/anthy/ticket/33075
BUG=https://groups.google.com/forum/#!topic/ibus-user/eRr0Ix4GbgE