UTF-8 as default encoding (essential change) (code)
Uninitialized variable such as encoding is 0, thus
ANTHY_UTF8_ENCODING ANTHY should use 0.

Although this problem was spotted for the code in test/*,
the most significant problem was the code in calctrans/*
which does statistics.

This problem was identified while trying to rebuild of
anthy.dep.  With this fix, first conversion candidate improved a lot.

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=928890

0.3-8.1 (Before)
世武藤よ位。
同様で素。
なくなるか名
なる歩度。
読むと良い。
同様で酢。
同様です。
同様です。
somehow some conversion improves but initial one is very bad.

0.3-9 (After)
読むとよい。
同様です。
Quite reasonable.

Signed-off-by: Osamu Aoki <osamu@debian.org>