Subject: Use UTF-8 for zipcode conversion (essential change) (code)
* src-worddic/ext_ent.c (pushback_place_name): Use UTF-8.

This is taken from a commit in the upstream's development branch
containing no auto-generated file (branch=master) of salsa(NIIBE Yutaka).
Somehow, this was not included in the branch=release where the upstream
released tarball contents was intended to be tracked (but ... that
doesn't seem to happen).

Original commit:
97f5ef0 ("Use UTF-8 for zipcode conversion.", 2015-04-28)

Space adjusted and dropped non-essential changes to ZIPDICT_EUC.
(I skipped all attempts to change ZIPDICT_EUC to ZIPDICT..
 That should be done after discussion with previous maintainers.
 This is intended to be a pure bug fix patch.)

Signed-off-by: Osamu Aoki <osamu@debian.org>