Convert depgraph/master.depword to UTF-8 (data)
This is very odd to see this data file content in non UTF-8 format.
Let's converted this to UTF-8.

This should change behavior of anthy at least for
  @接続詞 "" Sy@ Sy@間投助詞

Since other entries are mostly commented out, this damage was
relatively small.  EUC-JP and UTF-8 encoding mixed.  So this part
of depword is broken as it was.

Signed-off-by: Osamu Aoki <osamu@debian.org>